update: reorganize code from studio to a standalone converter package
