+++ converter/openaigeminiconverter/openai_to_gemini_test.go
@@ -1,4 +1,5 @@
-package studio
+
+package openaigeminiconverter
 
 import (
 	"encoding/json"
